refactor(cloudformation): move getOptTags to opt.go
